progress.c: call progress_interval() from progress_test_force_update()
To make it clear that test_progress_force_update() is a synonym for
progress_interval(), define for the former in terms of a single call
to the latter. Doing this without forward declarations requires moving
the code around, so let's move all of these above their respective
users.

In a subsequent commit this signal-like behavior of
"test_progress_force_update" will start to matter.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>