TODO: fix up synchronicity in say_color_tap
Without this I can reproduce a flaky run in this test:

    for i in {1..20}; do make -C t prove T=t5811-proto-disable-git.sh; done

This is because we get a git-daemon background process returning and
emitting a "Disconnected" line, which sometimes lands between the
prefix printf here and the actual TAP print.

So we absolutely *need* to insert these in a single write(), which I
think I'm guaranteed to do with a single printf:

    $ strace printf "%s%s\n" "hi" "there" 2>&1 | grep hi
    execve("/usr/bin/printf", ["printf", "%s%s\\n", "hi", "there"], 0x7ffc377f2188 /* 48 vars */) = 0
    openat(AT_FDCWD, "/usr/lib/locale/locale-archive", O_RDONLY|O_CLOEXEC) = 3
    write(1, "hithere\n", 8hithere

See [1] for details.

1. https://lore.kernel.org/git/CAHk-=wgyk0mwYcMRC8HakzoAKL2Y3gwzD433tqKYYhV+r1PLnA@mail.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>