fetch-negotiator init(): don't take a "struct repository"
Change fetch_negotiator_init() to initialize a backend we ask it to
initialize, instead of it taking a "struct repository *r" and doing
its own prepare_repo_settings(), add a fetch_pack_negotiator_init()
wrapper to fetch-pack.c to fit fetch-pack.c's current use case for
this API.

This makes it easier to switch negotiation strategies mid-request,
which I'm aiming to do further down the line.

It makes sense at this point to move the "enum fetch_negotiator_init"
to "fetch-negotiator.h" from "repository.h", and to have
"repository.h" include the former. We need to refer to the enum in the
"fetch_negotiator_init()" argument list.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>