Makefile: add and use shell quoting functions
Add functions to "shared.mak" which escape single quotes, the same
while adding surrounding single quotes etc. Then use this to define
"make-SQ-vars" and "make-CQ_SQ-vars" functions, by using these in the
Makefile we get rid of a lot of the line noise of defining these *_SQ
and *_CQ_SQ variables.

We had previously avoided doing this due to the "$(call)" feature of
GNU make being relatively new back when this was originally attempted
back in 4769948afe7 (Deal with $(bindir) and friends with
whitespaces., 2005-10-10). See 7ffe7098dca (Fix installation of
templates on ancient systems., 2006-07-29) for the commit which
reverted that. Due to those incompatibilities with older GNU make
versions, and we've used the more verbose *_SQ pattern ever since.

The "$(call)" feature was introduced in GNU make version 3.78,
released on the 22nd of September, 1999. That release also introduced
"$(error)" and "$(warning)", which we've been making use of since
f2fabbf76e4 (Teach Makefile to check header dependencies, 2010-01-26),
so we can safely use these.

This extends upon the macros added in 4769948afe7: We now have macros
for quoting a ' inside '', and a ' with no surrounding '' as before.

Additionally provide and use a "shqq" macro along with a "Cq" and
"shq_Csq" for the C strings being escaped here. These are used for the
common case of wanting to quote a C string we pass to the compiler
with a -D flag.

This doesn't get rid of all of our shell quoting. We've still got some
in the main Makefile, most notably in and around
GIT-BUILD-OPTIONS. Let's leave that for now, that case is a bit harder
to migrate, and that can be done some other time.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>